Ajout notification

+++ admin/supprimer-sous-categories.php
@@ -8,15 +8,15 @@ if (!(check_permission($conn, 'delete_subcategories'))) {
     $_SESSION['message-failed'] = "Vous n'avez pas la permission de voir cette page.";
     header("Location: admin.php");
     exit;
-  }
+}
 
 // Récupérer l'ID de la sous catégories à partir de la requête GET
 $subcategories_id = isset($_GET['id']) ? intval($_GET['id']) : 0;
 
 // Si aucun ID de sous catégories n'a été spécifié, rediriger l'utilisateur vers la page de gestion des sous-catégories
 if (!$subcategories_id) {
-   header("Location: gestion-sous-categories.php");
-   exit;
+    header("Location: gestion-sous-categories.php");
+    exit;
 }
 
 // Récupérer les informations sur l'utilisateur à partir de la base de données
@@ -26,34 +26,47 @@ $subcategories = mysqli_fetch_assoc($result);
 
 // Si la sous catégorie n'a pas été trouvé, redirige l'utilisateur vers la page de gestion des sous catégories
 if (!$subcategories) {
-   header("Location: gestion-sous-categories.php");
-   exit;
+    header("Location: gestion-sous-categories.php");
+    exit;
 }
 
 // Si le formulaire a été soumis
 if (isset($_POST['submit'])) {
+    // Vérifie s'il reste des articles dans la sous-catégorie
+    $sql = "SELECT COUNT(*) FROM articles WHERE subcategory_id = $subcategories_id";
+    $result = mysqli_query($conn, $sql);
+    $count = mysqli_fetch_row($result)[0];
+
+    if ($count > 0) {
+        $_SESSION['message-failed'] = "Il reste des articles dans cette sous-catégorie, vous ne pouvez pas la supprimer.";
+        header("Location: gestion-sous-categories.php");
+        exit;
+    }
+
     // Supprimer l'utilisateur de la base de données
     $sql = "DELETE FROM subcategories WHERE id = $subcategories_id";
     mysqli_query($conn, $sql);
 
     // Ajouter un message de réussite
-   $_SESSION['message-success'] = "La sous catégories à été supprimé avec succès.";
+    $_SESSION['message-success'] = "La sous catégories à été supprimé avec succès.";
 
     // Rediriger l'utilisateur vers la page de gestion des sous catégories
     header("Location: gestion-sous-categories.php");
     exit;
-    }
+}
 
-    ?>
+?>
 
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Thales - Supprimer une sous-catégories</title>
 </head>
+
 <body>
     <!-- Sidebar -->
     <?php include "navbar-admin.php"; ?>
@@ -70,4 +83,5 @@ if (isset($_POST['submit'])) {
         </form>
     </div>
 </body>
+
 </html>
